[Refactor] Move duplicated logic into shared file (#17)
* git ignore model checkpoints

* Move shared logic into shared.py
