+++ .gitignore
@@ -170,3 +170,6 @@ log
 # macOS
 .DS_Store
 Roms
+
+# model checkpoints
+*.pt
