+++ examples/nanoChatGPT/config/train.yaml
@@ -32,6 +32,6 @@ min_lr: 6.0e-5  # minimum learning rate, should be ~= learning_rate/10 per Chinc
 # DDP settings
 backend: nccl  # nccl, gloo, etc.
 # system settings
-device: cuda  # examples: cpu, cuda, cuda:0, cuda:1 etc., or try mps on macbooks
+device: cpu  # examples: cpu, cuda, cuda:0, cuda:1 etc., or try mps on macbooks
 dtype: bfloat16  # float32, bfloat16, or float16, the latter will auto implement a GradScaler
 compile: True  # use PyTorch 2.0 to compile the model to be faster
