+++ examples/nanoChatGPT/train.py
@@ -16,24 +16,28 @@ $ torchrun --nproc_per_node=8 --nnodes=2 --node_rank=1 --master_addr=123.456.123
 (If your cluster does not have Infiniband interconnect prepend NCCL_IB_DISABLE=1)
 """
 
-import math
 import os
-import pickle
 import time
-from contextlib import nullcontext
 from pathlib import Path
 from typing import Optional
 
 import numpy as np
 import torch
 import torch.nn as nn
-from nanoGPT.model import GPT, GPTConfig
 from tensordict.nn import TensorDictModule
 from tensordict.prototype import tensorclass
 from torch.distributed import destroy_process_group
 from torch.nn.parallel import DistributedDataParallel as DDP
-from torch.utils.data import DataLoader, Dataset
+from torch.utils.data import Dataset
 from utils import init_ddp, load_and_update_config
+
+from shared import (
+    create_infinite_dataloader,
+    create_lr_scheduler,
+    init_model,
+    load_checkpoint,
+    setup,
+)
 
 HERE = Path(__file__).parent
 
@@ -91,119 +95,15 @@ def create_datasets(config):
     return train_data, val_data
 
 
-def create_infinite_dataloader(data, config):
-    """
-    Creates a dataloader and yields batches from it indefinitely, so that we can request
-    batches whenever we like with next.
-    """
-    dl = DataLoader(
-        data,
-        batch_size=config["batch_size"],
-        shuffle=True,  # TODO: perhaps validation set shouldn't be shuffled?
-        collate_fn=Collate(config["device"]),
-        drop_last=True,
+def get_dataloaders(config):
+    train_data, val_data = create_datasets(config)
+
+    train_loader = create_infinite_dataloader(
+        train_data, config, Collate(config["device"])
     )
-    while True:
-        yield from dl
+    val_loader = create_infinite_dataloader(val_data, config, Collate(config["device"]))
 
-
-def load_checkpoint(config):
-    ckpt_path = os.path.join(config["out_dir"], "ckpt.pt")
-    return torch.load(ckpt_path, map_location=config["device"])
-
-
-def init_model_scratch(config, model_kwargs):
-    # attempt to derive vocab_size from the dataset
-    meta_path = HERE / "nanoGPT" / "data" / config["dataset"] / "meta.pkl"
-    meta_vocab_size = None
-    if os.path.exists(meta_path):
-        with open(meta_path, "rb") as f:
-            meta = pickle.load(f)
-        meta_vocab_size = meta["vocab_size"]
-        print(f"found vocab_size = {meta_vocab_size} (inside {meta_path})")
-    # init a new model from scratch
-    print("Initializing a new model from scratch")
-    # determine the vocab size we'll use for from-scratch training
-    if meta_vocab_size is None:
-        print(
-            "defaulting to vocab_size of GPT-2 to 50304 (50257 rounded up for efficiency)"
-        )
-    model_kwargs["vocab_size"] = (
-        meta_vocab_size if meta_vocab_size is not None else 50304
-    )
-    gptconf = GPTConfig(**model_kwargs)
-    return GPT(gptconf)
-
-
-def init_model_resume(config, model_kwargs):
-    print(f"Resuming training from {config['out_dir']}")
-    # resume training from a checkpoint.
-    checkpoint = load_checkpoint(config)
-    checkpoint_model_args = checkpoint["model_args"]
-    # force these config attributes to be equal otherwise we can't even resume training
-    # the rest of the attributes (e.g. dropout) can stay as desired from command line
-    for k in ["n_layer", "n_head", "n_embd", "block_size", "bias", "vocab_size"]:
-        model_kwargs[k] = checkpoint_model_args[k]
-    # create the model
-    gptconf = GPTConfig(**model_kwargs)
-    model = GPT(gptconf)
-    state_dict = checkpoint["model"]
-    # fix the keys of the state dictionary :(
-    # honestly no idea how checkpoints sometimes get this prefix, have to debug more
-    unwanted_prefix = "_orig_mod."
-    for k in state_dict:
-        if k.startswith(unwanted_prefix):
-            state_dict[k[len(unwanted_prefix) :]] = state_dict.pop(k)
-    model.load_state_dict(state_dict)
-    config["iter_num"] = checkpoint["iter_num"]
-    config["best_val_loss"] = checkpoint["best_val_loss"]
-    return model
-
-
-def init_model_gpt2(config, model_kwargs):
-    print(f"Initializing from OpenAI GPT-2 weights: {config['init_from']}")
-    # initialize from OpenAI GPT-2 weights
-    override_args = {"dropout": config["dropout"]}
-    model = GPT.from_pretrained(config["init_from"], override_args)
-    # read off the created config params, so we can store them into checkpoint correctly
-    for k in ["n_layer", "n_head", "n_embd", "block_size", "bias", "vocab_size"]:
-        model_kwargs[k] = getattr(model.config, k)
-
-    return model
-
-
-def init_model(config):
-    model_kwargs = {
-        "n_layer": config["n_layer"],
-        "n_head": config["n_head"],
-        "n_embd": config["n_embd"],
-        "block_size": config["block_size"],
-        "bias": config["bias"],
-        "vocab_size": None,
-        "dropout": config["dropout"],
-    }
-
-    if config["init_from"] == "scratch":
-        model = init_model_scratch(config, model_kwargs)
-    elif config["init_from"] == "resume":
-        model = init_model_resume(config, model_kwargs)
-    elif config["init_from"].startswith("gpt2"):
-        model = init_model_gpt2(config, model_kwargs)
-
-    # crop down the model block size if desired, using model surgery
-    if config["block_size"] < model.config.block_size:
-        model.crop_block_size(config["block_size"])
-        # so that the checkpoint will have the right value
-        model_kwargs["block_size"] = config["block_size"]
-    model.to(config["device"])
-
-    # compile the model
-    # TODO: is this fine here or do we need to init optimizer first?
-    if config["compile"]:
-        print("compiling the model... (takes a ~minute)")
-        model = torch.compile(model)  # requires PyTorch 2.0
-
-    return model, model_kwargs
+    return train_loader, val_loader
 
 
 def init_scaler(config):
@@ -211,13 +111,13 @@ def init_scaler(config):
     return torch.cuda.amp.GradScaler(enabled=(config["dtype"] == "float16"))
 
 
-def init_optimizer(config):
+def init_optimizer(model, config):
     # optimizer
     optimizer = model.configure_optimizers(
         config["weight_decay"],
         config["learning_rate"],
         (config["beta1"], config["beta2"]),
-        device_type,
+        "cuda" if "cuda" in config["device"] else "cpu",
     )
     if config["init_from"] == "resume":
         checkpoint = load_checkpoint(config)
@@ -240,36 +140,24 @@ def create_loss_estimator(config):
     return estimate_loss
 
 
-def create_lr_scheduler(config):
-    # learning rate decay scheduler (cosine with warmup)
-    def scheduler(it):
-        # 1) linear warmup for warmup_iters steps
-        if it < config["warmup_iters"]:
-            return config["learning_rate"] * it / config["warmup_iters"]
-        # 2) if it > lr_decay_iters, return min learning rate
-        if it > config["lr_decay_iters"]:
-            return config["min_lr"]
-        # 3) in between, use cosine decay down to min learning rate
-        decay_ratio = (it - config["warmup_iters"]) / (
-            config["lr_decay_iters"] - config["warmup_iters"]
-        )
-        assert 0 <= decay_ratio <= 1
-        coeff = 0.5 * (1.0 + math.cos(math.pi * decay_ratio))  # coeff ranges 0..1
-        return config["min_lr"] + coeff * (config["learning_rate"] - config["min_lr"])
+def train(config):
+    model, model_kwargs = init_model(config)
+    model.to(config["device"])
+    scaler = init_scaler(config)
+    optimizer = init_optimizer(model, config)
 
-    return scheduler
+    model = TensorDictModule(
+        model, in_keys=["prompt", "target"], out_keys=["logits", "loss"]
+    )
 
+    # compile the model
+    if config["compile"]:
+        print("compiling the model... (takes a ~minute)")
 
-def train(
-    model,
-    model_kwargs,
-    optimizer,
-    train_data,
-    val_data,
-    config,
-    master_process=True,
-    ddp=False,
-):
+        model = torch.compile(model)  # requires PyTorch 2.0
+    if config["is_ddp"]:
+        model = DDP(model, device_ids=[config["ddp_local_rank"]])
+
     # these will already have been set if resuming from previous checkpoint
     iter_num = config.setdefault("iter_num", 0)
     best_val_loss = config.setdefault("best_val_loss", 1e9)
@@ -283,15 +171,15 @@ def train(
 
     estimate_loss = create_loss_estimator(config)
 
-    train_loader = create_infinite_dataloader(train_data, config)
-    val_loader = create_infinite_dataloader(val_data, config)
+    train_loader, val_loader = get_dataloaders(config)
 
     # training loop
     next_batch = next(train_loader)  # fetch the very first batch
     t0 = time.time()
     local_iter_num = 0  # number of iterations in the lifetime of this process
-    # unwrap DDP container if needed, and unwrap TensorDictModule
-    raw_model = model.module.module if ddp else model.module
+    raw_model = (
+        model.module.module if config["is_ddp"] else model.module
+    )  # unwrap DDP container if needed
     running_mfu = -1.0
 
     while True:
@@ -301,7 +189,7 @@ def train(
             param_group["lr"] = lr
 
         # evaluate the loss on train/val sets and write checkpoints
-        if iter_num % config["eval_interval"] == 0 and master_process:
+        if iter_num % config["eval_interval"] == 0 and config["master_process"]:
             model.eval()
             losses = {
                 "train": estimate_loss(model, train_loader),
@@ -330,7 +218,7 @@ def train(
         # forward backward update, with optional gradient accumulation to simulate larger batch size
         # and using the GradScaler if data type is float16
         for micro_step in range(config["gradient_accumulation_steps"]):
-            if ddp:
+            if config["is_ddp"]:
                 # in DDP training we only need to sync gradients at the last micro step.
                 # the official way to do this is with model.no_sync() context manager, but
                 # I really dislike that this bloats the code and forces us to repeat code
@@ -359,7 +247,7 @@ def train(
         t1 = time.time()
         dt = t1 - t0
         t0 = t1
-        if iter_num % config["log_interval"] == 0 and master_process:
+        if iter_num % config["log_interval"] == 0 and config["master_process"]:
             # loss as float. note: this is a CPU-GPU sync point
             lossf = batch.loss.item()
             if local_iter_num >= 5:  # let the training loop settle a bit
@@ -383,49 +271,10 @@ def train(
 
 if __name__ == "__main__":
     config = load_and_update_config("config/train.yaml")
+    config.update(init_ddp(config["backend"], config["device"]))
 
-    ddp_config = init_ddp(config["backend"], config["device"])
+    ctx = setup(config)
+    train(config)
 
-    if ddp_config["master_process"]:
-        os.makedirs(config["out_dir"], exist_ok=True)
-    torch.manual_seed(1337 + ddp_config["seed_offset"])
-
-    torch.backends.cuda.matmul.allow_tf32 = True  # allow tf32 on matmul
-    torch.backends.cudnn.allow_tf32 = True  # allow tf32 on cudnn
-    device_type = (
-        "cuda" if "cuda" in config["device"] else "cpu"
-    )  # for later use in torch.autocast
-    # note: float16 data type will automatically use a GradScaler
-    ptdtype = getattr(torch, config["dtype"])
-    ctx = (
-        nullcontext()
-        if device_type == "cpu"
-        else torch.amp.autocast(device_type=device_type, dtype=ptdtype)
-    )
-
-    model, model_kwargs = init_model(config)
-    scaler = init_scaler(config)
-    optimizer = init_optimizer(config)
-
-    model = TensorDictModule(
-        model, in_keys=["prompt", "target"], out_keys=["logits", "loss"]
-    )
-    # wrap model into DDP container
-    if ddp_config["is_ddp"]:
-        model = DDP(model, device_ids=[ddp_config["ddp_local_rank"]])
-
-    train_data, val_data = create_datasets(config)
-
-    train(
-        model,
-        model_kwargs,
-        optimizer,
-        train_data,
-        val_data,
-        config,
-        master_process=ddp_config["master_process"],
-        ddp=ddp_config["is_ddp"],
-    )
-
-    if ddp_config["is_ddp"]:
+    if config["is_ddp"]:
         destroy_process_group()
